[Stdlib] Remove GPU print function and add support for sep/end (#33554)
This removes the GPU print function (which will be moved to a different
file in a subsequent PR). It removes the specialized unary overloads of
print, because they were only really useful for GPUs.

With these changes, we can actually have #32569 to work, so just do that
and support sep/end in the variadic print overload.

Closes #12811

modular-orig-commit: 8c911f6dbb3007f931e796a549a8e33ba27cf453

+++ stdlib/test/builtin/test_bfloat16.mojo
@@ -50,21 +50,25 @@ fn test_bf_primitives():
 def main():
     # CHECK: 33.0
     print(
-        __mlir_op.`pop.cast`[_type = __mlir_type[`!pop.scalar<f64>`]](
-            __mlir_op.`kgen.param.constant`[
-                _type = __mlir_type[`!pop.scalar<bf16>`],
-                value = __mlir_attr[`#pop.simd<"33"> : !pop.scalar<bf16>`],
-            ]()
+        Float64(
+            __mlir_op.`pop.cast`[_type = __mlir_type[`!pop.scalar<f64>`]](
+                __mlir_op.`kgen.param.constant`[
+                    _type = __mlir_type[`!pop.scalar<bf16>`],
+                    value = __mlir_attr[`#pop.simd<"33"> : !pop.scalar<bf16>`],
+                ]()
+            )
         )
     )
 
     # CHECK: nan
     print(
-        __mlir_op.`pop.cast`[_type = __mlir_type[`!pop.scalar<f64>`]](
-            __mlir_op.`kgen.param.constant`[
-                _type = __mlir_type[`!pop.scalar<bf16>`],
-                value = __mlir_attr[`#pop.simd<"nan"> : !pop.scalar<bf16>`],
-            ]()
+        Float64(
+            __mlir_op.`pop.cast`[_type = __mlir_type[`!pop.scalar<f64>`]](
+                __mlir_op.`kgen.param.constant`[
+                    _type = __mlir_type[`!pop.scalar<bf16>`],
+                    value = __mlir_attr[`#pop.simd<"nan"> : !pop.scalar<bf16>`],
+                ]()
+            )
         )
     )
